Add email and phone numbers

+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/SessionRealmConfigurationFactory.kt
@@ -19,13 +19,14 @@ package org.matrix.android.sdk.internal.database
 
 import android.content.Context
 import androidx.core.content.edit
+import io.realm.Realm
+import io.realm.RealmConfiguration
+import org.matrix.android.sdk.BuildConfig
 import org.matrix.android.sdk.internal.database.model.SessionRealmModule
 import org.matrix.android.sdk.internal.di.SessionFilesDirectory
 import org.matrix.android.sdk.internal.di.SessionId
 import org.matrix.android.sdk.internal.di.UserMd5
 import org.matrix.android.sdk.internal.session.SessionModule
-import io.realm.Realm
-import io.realm.RealmConfiguration
 import timber.log.Timber
 import java.io.File
 import javax.inject.Inject
@@ -46,20 +47,16 @@ internal class SessionRealmConfigurationFactory @Inject constructor(
         val migration: RealmSessionStoreMigration,
         context: Context) {
 
-    companion object {
-        const val SESSION_STORE_SCHEMA_VERSION = 3L
-    }
-
     // Keep legacy preferences name for compatibility reason
     private val sharedPreferences = context.getSharedPreferences("im.vector.matrix.android.realm", Context.MODE_PRIVATE)
 
     fun create(): RealmConfiguration {
         val shouldClearRealm = sharedPreferences.getBoolean("$REALM_SHOULD_CLEAR_FLAG_$sessionId", false)
         if (shouldClearRealm) {
-            Timber.v("************************************************************")
-            Timber.v("The realm file session was corrupted and couldn't be loaded.")
-            Timber.v("The file has been deleted to recover.")
-            Timber.v("************************************************************")
+            Timber.e("************************************************************")
+            Timber.e("The realm file session was corrupted and couldn't be loaded.")
+            Timber.e("The file has been deleted to recover.")
+            Timber.e("************************************************************")
             deleteRealmFiles()
         }
         sharedPreferences.edit {
@@ -74,7 +71,7 @@ internal class SessionRealmConfigurationFactory @Inject constructor(
                     realmKeysUtils.configureEncryption(this, SessionModule.getKeyAlias(userMd5))
                 }
                 .modules(SessionRealmModule())
-                .schemaVersion(SESSION_STORE_SCHEMA_VERSION)
+                .schemaVersion(RealmSessionStoreMigration.SESSION_STORE_SCHEMA_VERSION)
                 .migration(migration)
                 .build()
 
@@ -90,6 +87,11 @@ internal class SessionRealmConfigurationFactory @Inject constructor(
 
     // Delete all the realm files of the session
     private fun deleteRealmFiles() {
+        if (BuildConfig.DEBUG) {
+            Timber.e("No op because it is a debug build")
+            return
+        }
+
         listOf(REALM_NAME, "$REALM_NAME.lock", "$REALM_NAME.note", "$REALM_NAME.management").forEach { file ->
             try {
                 File(directory, file).deleteRecursively()
